Move RemoveProfile to other package
